Validator add

+++ app/Http/Controllers/Country/CountryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Country;
 use App\Http\Controllers\Controller;
 use App\Models\CountryModel;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Validator;
 
 class CountryController extends Controller
 {
@@ -15,20 +16,42 @@ class CountryController extends Controller
 
     public function countryById($id)
     {
-        return response()->json(CountryModel::find($id),200);
+        $country= CountryModel::find($id);
+        if (is_null($country))
+        {
+            return response()->json(["message"=>"Record Not Found!"], 404);
+        }
+        return response()->json($country,200);
     }
     public function countrySave (Request $request)
     {
+        $rules = [
+            'name' => 'required|min:3',
+        ];
+        $validator = Validator::make($request->all(), $rules);
+
+        if($validator->fails())
+        {
+            return response()->json($validator->errors(),400);
+        }
         $country = CountryModel::create($request->all());
         return response()->json($country,201);
     }
-    public function countryUpdate(Request $request, CountryModel $country)
+    public function countryUpdate(Request $request, $id)
     {
+        $country = CountryModel::find($id);
+        if (is_null($country)){
+            return response()->json(["message"=>"Record Not Found!"], 404);
+        }
         $country->update($request->all());
         return response()->json($country,200);
     }
-    public function countryDelete(Request $request, CountryModel $country)
+    public function countryDelete(Request $request, $id)
     {
+        $country = CountryModel::find($id);
+        if (is_null($country)){
+            return response()->json(["message"=>"Record Not Found!"], 404);
+        }
         $country->delete();
         return response()->json(null, 204);
     }
